Extract location code into it's own file
